+++ src/routes/Feed/Feed.js
@@ -1,14 +1,13 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { get, pick } from 'lodash/fp'
-
-import './Feed.scss'
+import { bgImageStyle } from 'util/index'
 import FeedList from 'components/FeedList'
 import Loading from 'components/Loading'
 import FeedBanner from 'components/FeedBanner'
 import TopicFeedHeader from 'components/TopicFeedHeader'
 import Button from 'components/Button'
-import { bgImageStyle } from 'util/index'
+import './Feed.scss'
 
 export default class Feed extends Component {
   static propTypes = {
@@ -59,28 +58,31 @@ export default class Feed extends Component {
 
   render () {
     const {
-      routeParams, topic, group, currentUser, postsTotal, followersTotal,
+      routeParams, group, currentUser, postsTotal, followersTotal,
       groupTopic, newPost, currentUserHasMemberships,
-      goToCreateGroup, membershipsPending, postTypeFilter
+      goToCreateGroup, membershipsPending, postTypeFilter, topicLoading, toggleGroupTopicSubscribe
     } = this.props
     const { topicName, context } = routeParams
 
-    if (topicName && !topic) return <Loading />
-    if (group && topicName && !groupTopic) return <Loading />
+    if (topicLoading) return <Loading />
     if (!currentUser) return <Loading />
     if (membershipsPending) return <Loading />
 
     return <div>
       {topicName
         ? <TopicFeedHeader
-          groupTopic={groupTopic}
+          isSubscribed={groupTopic && groupTopic.isSubscribed}
+          toggleSubscribe={
+            groupTopic
+              ? () => toggleGroupTopicSubscribe(groupTopic)
+              : null
+          }
           topicName={topicName}
           postsTotal={postsTotal}
           followersTotal={followersTotal}
-          topic={topic}
           type={postTypeFilter}
-          group={group}
           currentUser={currentUser}
+          bannerUrl={group && group.bannerUrl}
           newPost={newPost} />
         : <FeedBanner
           group={group}