+++ src/routes/PrimaryLayout/PrimaryLayout.js
@@ -11,7 +11,6 @@ import qs from 'querystring'
 import Intercom from 'react-intercom'
 import config, { isTest } from 'config'
 import AddLocation from 'routes/Signup/AddLocation'
-import AddSkills from 'routes/Signup/AddSkills'
 import AllTopics from 'routes/AllTopics'
 import CreateGroup from 'components/CreateGroup'
 import CreateModal from 'components/CreateModal'
@@ -102,7 +101,6 @@ const createRoutes = [
 const signupRoutes = [
   { path: '/signup/upload-photo', child: UploadPhoto },
   { path: '/signup/add-location', child: AddLocation },
-  { path: '/signup/add-skills', child: AddSkills },
   { path: '/signup/review', child: Review }
 ]