+++ src/routes/UserSettings/UserGroupsTab/UserGroupsTab.connector.js
@@ -2,7 +2,6 @@ import get from 'lodash/get'
 import { connect } from 'react-redux'
 import { createAffiliation, deleteAffiliation, leaveGroup } from './UserGroupsTab.store'
 import { createSelector as ormCreateSelector } from 'redux-orm'
-import fetchForCurrentUser from 'store/actions/fetchForCurrentUser'
 import getCurrentUserMemberships from 'store/selectors/getCurrentUserMemberships'
 import orm from 'store/models'
 
@@ -27,7 +26,6 @@ export function mapStateToProps (state, props) {
 
 export function mapDispatchToProps (dispatch) {
   return {
-    fetchForCurrentUser: (params) => dispatch(fetchForCurrentUser(params)),
     createAffiliation: (params) => dispatch(createAffiliation(params)),
     deleteAffiliation: (params) => dispatch(deleteAffiliation(params)),
     leaveGroup: (params) => dispatch(leaveGroup(params))
@@ -38,8 +36,7 @@ export function mergeProps (stateProps, dispatchProps, ownProps) {
   return {
     ...ownProps,
     ...stateProps,
-    ...dispatchProps,
-    fetchForCurrentUser: () => dispatchProps.fetchForCurrentUser(ownProps.personId)
+    ...dispatchProps
   }
 }